requirements.txt
bs4==0.0.2
faiss-cpu==1.11.0
numpy==2.3.1
openai==1.91.0
pandas==2.3.0
requests==2.32.4
sentence-transformers==4.1.0
streamlit==1.46.0
